Switches bridge dependency to official mono recipe.
Homebrew seems to have accepted a mono setup with msbuild included.

The barely maintained bridgedotnet/cli/mono's bottle was still for high_sierra
and was being poured when installed from mojave, which was resulting in an
unusable Bridge setup.

Pointing to the official homebrew package fixes the issue.

Standalone mono installation is not affected by this issue.